Populate SchematicsToDisable fully (#888)
Not all core LWOTC schematics were included in the `SchematicsToDisable`
config array.

In addition, that array is being used to clear the `CreatorTemplateName`
property on items, which needs to happen to prevent items automatically
being upgraded when awarded to the player. So I added the
'AdvancedGrenades' project to the list, even though it's not a
schematic, because the Mk2 grenades have it as the creator template.
This resulted in Mk1 grenades being upgraded if the player had
researched AdvancedGrenades, but not necessarily the individual Bomb
projects.

Fixes #888