Attempt to fix issues with rescuing captured soldiers
Flying blind here, but I noticed a couple of issues in the code:

 1. Some of the missions weren't checking in `NewGameState` to see
    whether a captured soldier was a reward already or not

 2. Rescued soldiers weren't being removed from
    `AlienHQ.CapturedSoldiers` in the case where they were captured, but
    no by a Chosen.

The first case has been fixed by passing a `NewGameState` to the code
that checks for the existence of captured soldier rewards on missions
and covert actions. The second has been fixed by removing soldiers from
that array when the soldier is given as a reward.

Note that there is still potentially an issue if a soldier dies during
the rescue mission.