Make Mission Screens hotlink compatible for controllers and keyboard compatible for M&K users.
Modifies : UIMission_LWCustomMission

The file has been re-structured so that functions appear, generally speaking, in the order in which they are called; this makes 'following the code' much easier.

------------------------------

Converts MissionInfoText from a UITextContainer to a UIVerticalScrollingText2

MissionInfoText is used to display LW specific mission information on the bottom of the mission panel, located on the left side of the screen. The problem with UITextContainers is that they don't display scrolled text; this appears to be an inherent bug which was fixed by LW2 coders via the creation of UIVerticalScrollingText2.

------------------------------

Mission screens display hotlinks for controller users and nice, wide conventional buttons for mouse and keyboard users.

BindLibraryItem was modified such that, when using a controller :
1.] Buttons are given the style eUIButtonStyle_HOTLINK_BUTTON; this way they appear as hotlinks.
2.] Buttons are given a gamepad icon so the hotlink appears properly.
3.] Buttons are set to resize to their text size; this helps center them within their parent panel.

BindLibraryItem was modified such that, when using a mouse and keyboard :
1.] Buttons don't resize to their text size; this way their position and size can be set manually.

* Similar changes were also made in BuildAlienFacilityOptionsPanel and BuildGoldenPathOptionsPanel for LockedButton; this is a button which appears when a mission is 'locked'.

** In order to center hotlinks within their parent panel, they call OnUnlockedButtonSizeRealized once realized; this is necessary since their text width is not initially known. LockedButton, rather than calling OnUnlockedButtonSizeRealized, calls OnLockedButtonSizeRealized.

------------------------------

The code within BuildSupplyRaidMissionPanel and BuildSupplyRaidOptionsPanel has been swapped since each had 'the others' code; the same was done for BuildLandedUFOMissionPanel and BuildLandedUFOOptionsPanel.

------------------------------

Since UIMission.RefreshNavigation is a mess; RefreshNavigation was overridden. It performs the following steps :
1.] It kills the navigation system and current focus.
2.] If using a mouse and keyboard, it adds visible buttons to the Navigator and selects the 1st appropriate one. If using a controller it does nothing, since you shouldn't be able to 'navigate' hotlinks.

Ultimately, this removes navigation for controller users who use hotlinks instead, and allows mouse and keyboard users to make use of the arrow keys to navigate visible buttons.

------------------------------

OnUnrealCommand was overridden so that mouse and keyboard users can use the spacebar or enter key to 'activate' the selected button.

------------------------------

Miscellaneous unnecessary code was removed; this is described within the code's comments.

******************************

Modifies : UIUtilities_LW.uc

Added the function AddBtnToNavigatorAndSelect which adds a UIButton, if it exists and is visible, to a UIScreen's Navigator; furthermore, it selects the UIButton if SelectionSet was false beforehand.

******************************

Adds : UIScreenListener_Mission_GoldenPath.uc

A screen listener which enables Golden path mission screens to follow the precedent set in custom LW mission screens (UIMission_LWCustomMission).
- Controller users see parent-panel centered hotlinks.
- Mouse and keyboard users see parent-panel centered conventional buttons which are navigable via the arrow keys.

******************************

Adds : UIScreenListener_Mission_ChosenStronghold.uc

A screen listener which enables the Chosen Stronghold mission screen to follow the precedent set in custom LW mission screens.
- Controller users see parent-panel centered hotlinks.
- Mouse and keyboard users see parent-panel centered conventional buttons which are navigable via the arrow keys.

******************************

Adds : UIScreenListener_Mission_ChosenAvengerAssault.uc

A screen listener which enables the Chosen Avenger Assault mission screen to follow the precedent set in custom LW mission screens.
- Controller users see parent-panel centered hotlinks.
- Mouse and keyboard users see parent-panel centered conventional buttons which are navigable via the arrow keys.

******************************

Adds : UIScreenListener_UFOAttack.uc

A screen listener which enables the UFO Attack mission screen to follow the precedent set in custom LW mission screens.
- Controller users see parent-panel centered hotlinks.
- Mouse and keyboard users see parent-panel centered conventional buttons which are navigable via the arrow keys.

******************************

Adds : UIScreenListener_Mission_GPIntelOptions.uc

A screen listener which enables the final mission screens to follow the precedent set in custom LW mission screens.

In this case, the navigation system was not modified when the mission is 'unlocked' since the UI is more complex; it consists of a list of intel-related UIButtons as well as normal 'accept' and 'cancel' UIButtons.