Fix alien customisation for Chosen summons
Alien customisation wasn't being applied in the case of Chosen summons,
so units like grenadiers and gunners were appearing as floating heads.

The customisation is now applied via an 'OnUnitBeginPlay' listener that
creates a new game state to create the customisation component. That
game state has an associated build visualisation function that renders
the customisation.

The XComGameStateContext_ChangeContainer used to create the
customisation component uses an `AssociatedPlayTiming` of
`SPT_BeforeSequential` to ensure that it occurs before the spawned units
are made visible.

This change also seems to render the `LWUpdateAlienCustomizationAction`
hack no longer necessary, so it has been removed. I have also moved all
the important code from `XCGS_AlienCustomizationManager` to a new event
listener class, since that's all the game state class was used for.

This also means that I have removed the hack for issue #117, since the
new listener resolves the customisation of reinforcement units as well.