Make Bring em On work with explosives
This updates the localisation for the ability to reflect the change.

This also fixes Bring em On for Technicals by assigning it to the
primary weapon slot.

In addition, I have added an `<Ability:BOUND_WEAPON_NAME/>` tag that
evaluates to the name of the weapon that is in the slot that the ability
is bound to. This means the localisation text works for both primary and
secondary weapons, or even pistols.