Adjust reinforcements for timer modifiers
Some sit reps adjust the mission timer, typically to add a few turns
when the other effects they add make the mission take longer. This
change modifies the continuous reinforcements system so that it takes
account of such timer modifiers and speeds them up or slows them down
accordingly.

As part of this work, I moved the large-map and very-large-map timer
bonuses to sit reps so that they work like other sit reps that modify
the timer. This has an advantage that players will know in advance that
they will be playing on a particularly large map or not, although it
only applies to fixed-evac missions with an evac timer.

In addition, I had to rework how timer modifications work, because
`X2SitRepEffect_ModifyKismetVariable`, which is used by the existing sit
rep effects like 'IncreaseTimer2Effect', can't stack with itself when
modifying the same Kismet variable. In other words, if a mission has two
sit reps that modify the Timer.LengthDelta Kismet variable, only one of
them "wins" - their values can't stack.

As a solution, my new IncreaseTimerNEffect_LW effects are of a new type
added by the Community Highlander that add a new game state object to
the mission start state that stores the current timer modifier. Each new
effect adds to or subtracts from that timer modifier. This allows them
to stack without issue. `SeqAct_InitializeMissionTimer` now fetches the
latest game state object of the appropriate type and uses its stored
modifier to update the mission timer.