Stabilise reinforcements for ambush missions
The reinforcements depended on both the region strength where the covert
action was run (which is often uncontacted by the player and may have a
lot of strength) and the liberation status of the region (reinforcements
weren't dropping in that case).

This fixes the effective region strength so that there isn't too much
variability in how quickly reinforcements arrive, and the mission is
hard-coded to include reinforcements even in liberated regions.