Fix numeric values in Apex Predator and Firestorm descriptions.
For some reason numeric values of APEX_PREDATOR_PANIC_RADIUS and FIRESTORM_DAMAGE_BONUS are displayed as float numbers (4.0000 and 2.0000 respectively).
I failed to understand why it is happening in these particular cases and decided to simply forcibly convert those to integer values for the descriptions. If there is less hacky solution for this, then it probably should be used instead.