Change priority of Flashbanger and Smoker templates
Currently the interaction between Smoker/Flashbanger and Full Kit abilities depend on the order of which they were acquired. So, Full Kit is inconsistent in granting or not extra flashbang/smoke grenade. Additionally, per Full Kit description it is arguable that it should not apply to these extra grenades at all.
This changes priorities of Smoker/Flashbanger templates, so that they happen after Full Kit and therefore would not have extra charges.