Improving Technical abilities descriptions
Changed descriptions for a number of Technical's abilities.

Formidable:  make use of config values for extra ablative armor and explosive damage reduction. Note, that for explosive damage reduction description  previouslyused vanilla <BlastPadding> tag, which show the value of BLAST_PADDING_DMG_ADJUST config setting. While it has same value of 50% as the FORMIDABLE_EXPLOSIVES_DR it is *not* what is used as a source for the ability.

Concussion Rocket: mention that allied units are immune to damage from it, disorient chance does *not* depend on gauntlet tier (unlike stun chance, and maybe unintended), mention that there are units (PsiZombie, AdvPsiWitchM2 and AdvPsiWitchM3) which are explicitly immune to all of the effects. Note, that the disorient effect is created with"mental" damage type for some weird reason and is probably unintended and should be fixed elsewhere. Otherwise, it should be added to the description as well.

Burnout: make use of config values for smoke cloud radius and default smoke defense bonus. Surprisinly to me, I have not found vanilla description tag for SMOKEGRENADE_HITMOD so I added one myself.

Napalm-X: fix to a small error in previously added End Translation tag.