Set Barrier sitrep bNegativeEffect flag to true
DefaultGameData does not set Barrier dark event sitrep as negative one, while it clearly is, which leards to its description shown in white text instead of red. This fixes it.
Also, DefaultGameData adds Barrier effect template to PositiveEffects array instead of NegativeEffects one, but it seems that there is no practical differences between the two, so it is fine to leave it as is.