Move some abilities assignments from characters to weapons (config version) (#1259)
This does the same thing as the previous PR "Move some abilities assignments from characters to weapons" did but through different method. "LockedOn", "Executioner_LW" and "CloseAndPersonal" were excluded from the change as it was not necessary for the desctiptions of those. Part of previous comment (below) still applies.

Changed "CloseCombatSpecialist", "BringEmOn" and "CoupdeGrace2" from being assigned character templates to being assigned to weapon templates.

For "CloseCombatSpecialist", "CoupdeGrace2" and "BringEmOn" it is required for Ability:BOUND_WEAPON_NAME tag in the descriptions to work correctly. Additionally, as far as I can tell BringEmOn didn't even work properly the way it was set up.

Co-authored-by: Grobobobo <50756826+Grobobobo@users.noreply.github.com>