Adding WOTC enemies to couple of lists they were missing from (#1249)
Added WOTC enemies to the list, which make them not drop corpses when killed by explosions. It is rather pointless for the Lost currently, but better to ass it just in case of future changes.

Added Purifiers and Priests to be eligible for Mutons' Warcry ability. Spectres are left out, because they are robotic (ability itself excludes robotic targets anyway). Credit to Anthogator for pointing out missing entries in Muton Warcry list a while back.

As a sidenote, I think doing both of these lists the way they are done is probably not particularly friendly to other enemy mods. Especially first one, since it is done in script. I wonder if it would be better to move to default behaviour with configurable exclusion lists, but I don't care much about other mods, so I will leave it as is.