Make faceless-recruit-chance modifier abilities configurable
Players and mods can now override which abilities contribute to the
reduction in chance to recruit faceless in havens. This is done via the
new `FACELESS_CHANCE_REDUCTION_ABILITIES` config array.

In addition, the contribution of successive abilities diminishes the
more abilities the liaison has. So having two of the abilities has a
significant impact, but 4 or more isn't much better than 2.