Fix Fire and Steel not applying to burn tick damage
Fire and Steel currently do not apply its bonus damage to burn tick damage, despite the description saying it should. The issue and the solution are the same as what was solved in "Make Boosted Cores apply to tick damage".

Since burning effect have bIgnoreBaseDamage flag set, the normal X2Effect_BonusWeaponDamage won't apply to it. X2Effect_BonusWeaponDOT copies the base functionality of X2Effect_BonusWeaponDamage with an option for it to apply to sources with bIgnoreBaseDamage set, such as DOT effects.