Prevent Field Surgeon to apply to bleeding out soldiers
Current description of Field Surgeon says: "It does not affect soldiers who suffered killing wounds and are stabilized from bleeding out.". However, it is erroneus for several reasons:
1. The code in X2Effect_FieldSurgeon checks only for BleedingOut status and not for Unconscious for target soldier.
2. BleedingOut status would be explicitly cleared prior to the Field Surgeon checks, if the soldier was carried out by someone else during evacuation (which is often the case).
3. Most importantly, both BleedingOut and Unconscious effects are cleared when the unit leaves play, which is way before Field Surgeon checks happen.

This changes X2Effect_FieldSurgeon effect to store the fact that soldier was bleeding out during any point of the mission and use this knowledge later to check whether Field Surgeon should apply.

To note. there are checks in the code for whether the specialist himself is unconscious/bleeding out. It supposed to prevent Field Surgeon to apply in these cases. However, for the same reasons these checks do nothing. This is not fixed in this commit. It might be okay, since Field Surgeon description does not mention this requirement in the first place, however it can lead to somewhat strange behavior where if the specialist himself was bleeding out he will heal everyone else, but not himself.