Allow different weapon categories in the pistol slot
Players and mods can now use the new `PISTOL_SLOT_WEAPON_CATS` config
variable to expand the number of weapons allowed in the pistol slot.
This array is empty by default so that there is an option to remove the
"pistol" weapon category if so desired. If the array is empty, a default
collection of weapon categories is used instead (currently it consists
solely of the "pistol" weapon cat).

Another change is that mods or players can now disable the pistol slot
via the new DISABLE_LW_PISTOL_SLOT config var, allowing them to safely
use mods like Dedicated Pistol Slot with LWOTC.

Lastly, I changed the way that PistolStandardShot is patched to support
Quickdraw. The new approach should work better with any mods that add or
modify the ability costs for that ability.