Update CHItemSlot_PistolSlot_LW.uc
I did not edit the check for allowing the item to show on the Inventory UI (so whilst soldiers could equip the specified items, they would not show in the Armory UI). I confirmed this by having a Templar with an Autopistol equipped on the secondary despite those being in eINVSlot_Pistol (Autopistol Overhaul allows sidearms and pistols to go through both slots.