Fix float-heads bug for reinforcements
The `OnUnitBeginPlay` listener responsible for customising the LW2 units
was calling `ModifyStateObject()` on the game state passed to the
listener, not the new game state created specifically for the unit
customisations. Derp.

This should at least fix the issue for reinforcements. Don't know if it
fixes the problem with loading tactical saves.