Rework Templar and rebalance other hero classes
First, the Templar now has a ballistic shield in place of the autopistol
as the secondary weapons, which also means the class has a one-handed
gauntlet rather than a pair. The ability tree has been completely
reworked to account for the shield, with three broad themes of Tank,
Melee Damage Dealer and Caster.

Another important change to Templar is that Rend now grants focus on any
attack, not just kills. Even a miss will grant focus.

Second, hero classes can no longer purchase multiple class abilities at
each rank by default. This brings them into line with the other soldier
classes. The second wave option for allowing such purchases now applies
to hero classes as well.

Because of this nerf to the late-game potential of the hero classes,
the trees for the Reaper and Skirmisher have been updated. In addition,
Reaper's Shadow now lasts 2 turns, but Shadow Killer no longer has a
chance to extend the duration.

As the autopistol is no longer the secondary for the Templar, it can now
be used by any soldier class in the pistol slot. This also means the
Templar now has the pistol slot too. It has shorter range and less base
damage than the pistol, but more crit damage and a lot more crit chance.