Apply Insider Knowledge continent bonus to all weapon mods (#288)
Currently Insider Knowledge continent bonus only applies to Laser Sights, Auto-Loaders and Expanded Magazines (because they are functionally unchanged from vanilla). This changes the other weapon mods (Scopes,, Hair Triggers, Stocks and Suppressors) so that Insider Knowledge improves them too, like it says it should.

It may be argued that these weapon mods or this continent bonus do not need buffs, so the bonus values I chose are small and might be overall insignificant.  But the principle of continent bonus actually working is still there and at least bonus values could be changed later on, if one would want to.

Fixes #288.