Make Disable Tutorial SWO work
Very few of the tutorial boxes were checking whether the tutorial was
disabled or not, so players were seeing them even when they had the
tutorial disabled. This fixes that.

I have also fixed the troop column tutorial box so that it only displays
the Gatecrasher text when the mission is in fact Gatecrasher, rather
than for all troop column missions.

Lastly, I've updated the x2proj file for all the new tutorial classes.