Prevent defeated Chosen appearing on missions
Since a Chosen can be attached to a mission on spawn and then defeated
before that mission is launched, we need to check whether the Chosen has
been defeated before adding them to the XComHQ tactical gameplay tags.
This should fix the issue with Chosen appearing on missions after being
defeated.