Improve Gatecrasher tutorials
This fixes an issue where the tutorials for Sentries, Gunners and
Engineers weren't appearing because their objectives depended on the
drone-sighting objective being completed first. This was because I was
using objective steps which need to be completed in order.

I took this opportunity to merge the separate tutorial boxes for the
infantry units into a single one. I also updated the text of all of the
Gatecrasher tutorial boxes to be less specific, while still (hopefully)
providing useful info.