Fix PCS behaviour with `RespecSelectedSoldier`
Apparently respeccing a soldier while they have a PCS installed makes
the soldier lose whatever bonus stat they were getting from the PCS,
while leaving the PCS in place. In effect, the soldier gets a permanent
stat penalty based on the PCS.

I've updated the command so that it removes the PCS from the soldier
before respeccing and then restores it afterwards.