Prevent continent bonus activation pre Resistance Radio
I have added a listener for the 'ContinentBonusActivated' event that
deactivates the continent bonus again if Resistance Radio has not been
researched yet. To ensure that it's activated again once the research
has been completed, I've also add an `ResearchCompletedFn` hook for
Resistance Radio that enables any eligible continent bonuses that were
blocked.

Lastly, I added a screen listener for the "continent bonus activated"
alert so that it's closed immediately (so the user doesn't see it) if
Resistance Radio hasn't yet been researched.

This is all to make sure that campaign starts on a 2-region continent
don't come with a *very* early continent bonus, since by default they
activate even if Resistance Radio hasn't been researched (this is
because LW2/LWOTC starts with a radio relay in the first haven, unlike
vanilla).