Fix crashes when Lost spawn
The mod shader cache we generate for our copy of ObelixDk's Lost Brute
appears to be missing something, because it appears to be the cause of
occasional crashes when a Lost Brute spawns.

This implements a workaround where we use a scaled up Howler model for
the Brute unless World War Lost is installed, in which case we use its
Brute model.