Fix Immune flyover on Solaced robotic units
The vanilla version of Solace ignores the multitarget conditions on the
ability, meaning that robotic units are granted immunity to mental,
disoriented, etc. *damage types* inflicted by an apply-weapon-damage
effect. That's why throwing a Bluescreen Bomb flashbang at a MEC in a
Solace aura would trigger an Immune flyover, even when the disorient
effect itself was applied properly.

I have replaced the vanilla effect with a subclass that adds a check on
the ability's multitarget conditions and excludes any units that don't
satisfy all of them