Modify pistol abilities to require pistol/sidearm
I have added a shooter condition to the pistol ability templates that
disables the abilities if the unit with those abilities does not have an
appropriate weapon equipped. The default requirement is that the soldier
have a pistol or sidearm, but other weapon categories can be added to
the new `PISTOL_ABILITY_WEAPON_CATS` config option in
XComLW_Overhaul.ini.