Add new weapon category for Templar shields
Using the "shield" category for Templar shields and partially
implementing the features of Ballistic Shields mod turned out not to
help compatibility much.

This change switches Templar shields to a different weapon category,
"templarshield", and only provides animation sets for weapons of that
category. This should make LWOTC compatible with the Ballistic Shields
mod. The downside is that there are now different shield types that look
different. Hopefully someone will provide an alternative model for the
Templar shield in the future.