Fix One For All bonus ablative for M2 and M3 shields
One For All is associated with the secondary weapon slot in
XComClassData.ini, but this association was lost in the effect's
implementation. The underlying problem is that One For All is overriding
Hunker Down. When one ability overrides another, you have to factor in
whether it should also override the associated weapon/item.

The default behaviour is for the overriding ability to be associated
with whichever weapon the overridden ability is tied to. Since Hunker
Down has no source weapon, One For All can't check the source weapon.

The fix is to set the One For All template's `bOverrideWeapon` property
to `true`.