Kill bleeding-out soldiers on ambushes (#1304)
The Covert Escape (ambush) mission definition was copied from WOTC and
for some reason it has triad objective set to true. LWOTC treats this as
meaning no enemy corpse retrieval, but XCOM and rebels are rescued if
the mission is successfully completed (think retaliations and network
tower).

We don't want soldiers automatically being rescued after completion of
the ambush, so I've set the triad objective status to false.

Fixes #1304.