Add Gatecrasher tutorial text for all languages
The first tutorial box on Gatecrasher was showing errors rather than
English for non-English installations. This was probably due to how
localised text arrays are handled. I've added the required entries (in
English) to all the other localisation files, so at least the proper
text will be displayed rather than an error. The hope is that
translations will be provided at a later date.