Fix incompatibility with CHL wrt failing CAs
The Community Highlander has made the new `RewardsNotGivenOnCompletion`
property on `XCGS_CovertAction` private write, which means we can no
longer set it directly to mark the covert action as having failed. To be
honest, that's probably a good thing as our existing solution was a bit
of a hack.

This commit introduces a single game state object,
`XCGS_CovertActionTracker`, that keeps track of whether a covert action
ambush mission was successful or not. This tracker is then used
alongside `RewardsNotGivenOnCompletion` to determine whether a covert
action failed or not, and hence whether the rewards should be given.