Fix enemies attacking units from yellow alert (#879)
This applies to Najas and Berserkers on Intel Raid missions, who can
either attack XCOM (Najas) or their squad mates (Berserkers) when
they're on yellow alert. The fix takes advantage of a new behaviour tree
node added to the Community Highlander, which allows us to restrict
targeting to just the priority target. The old behaviour node had a
fallback to other targets if the priority target could not be attacked
for any reason.