Lower alert level for early intel raids
While an interesting challenge, early intel raids were quite a way out
of line with other missions at this stage of the campaign. And I don't
think we need to make the early game necessarily harder.

This change lowers the alert for intel raids at FL4 and lower by 2, and
by 1 for FL5-8.