Update effects for CHL damage modifier hooks
This makes all %-based abilities in LWOTC use the new Community
Highlander hooks that ensure they work reliably, applying percentage
changes to all flat damage bonuses and penalties as well as being
independent of effect order.

I also updated `X2Effect_ApplyAltWeaponDamage` so that it's inline with
the latest CHL version of the base class. Ideally we would add an
overridable function to the CHL base class that returns the base weapon
damage. Then we could remove most of the code, but not ready to make
another highlander change at this point.