Migrate to X2ModCommonBuild
The old build has been replaced with one based on the community
X2ModCommonBuild, which has several advantages:

 * You no longer need to copy custom highlander code to SrcOrig
 * Assets can now be cooked
 * The build is faster when code or assets have not been changed

As part of the migration, the community highlander is now included as a
submodule, which is pointed at the `lwotc-dev` branch.

Mod Buddy should automatically pick up the new build changes, meaning
you won't need to do anything different to build the project in Mod
Buddy.

I have added various VS Code tasks so that LWOTC, the highlander, and
the DLC2 highlander can all be built from the LWOTC project.

Lastly, I added a new build script that can be used by translaters to
"install" the LWOTC project's localisation files into the installed mod,
converting them from UTF-8 to UTF-16 in the process. This build script,
which is accessible either from the new install-loc.bat batch file or a
VS Code task, does not require the WOTC SDK to run.