Fix performance issue with multi-shot abilities (#1340)
The follow-up shot abilities on multi-shot abilities were treating all
enemies on the map as valid targets, which was causing a lot of
unnecessary calculations. I have introduced a couple of solutions that
players can toggle between:

 1. Follow-up shots are only available if a unit value has been set by
 the first shot
 2. Follow-up shots are configured to require gameplay visibility

The first one has the downside that using the abilities still lags the
game somewhat. The second option means that follow-up shots may not fire
if the attacker loses line of sight on the target after a previous shot,
for example if cover providing step out is destroyed.

Players can switch between these two options for testing via the new
`USE_LOS_FOR_MULTI_SHOT_ABILITIES` config setting. If `true`, the second
option is used.