Tweak behaviour of Shadow
Shadow can now be activated even if the Reaper is flanked, but the
Shadow bonuses are no longer active when the Reaper starts a mission in
normal concealment.

I have also added `X2Effect_BreakSquadConcealment`, which isn't
currently used (I was testing out different behaviours), but it's a
non-trivial-enough effect to keep in case there's a use for it later on.