Adds a button on the outpost management screen to edit haven advisor loadout
Currently, the behavior when clicked via the avenger "Resistance Management" shortcut is a bit finnicky, in that backing out after editing the loadout results in two blank/unpopulated screens and then dumps you back in the barracks. I suspect the screens are still in the stack but for not showing for some reason. In any case, I plan to only make this button available when viewed through the geoscape (using a separate button), where this issue does not occur