Fix free abilities not showing as green icons
Many of the abilities that could be modified to be free, for example
flamethrower with Quickburn, are turn ending by default. However, the
turn-ending status was taking precedence over the free-action status for
icon colouring.

Now the turn-ending status is set to false if an ability is changed to a
free action.