Persist second wave options between campaigns
When a new campaign is started, or the second wave options are changed
mid campaign, the second wave options are now saved to a new INI file
(XComLW_SecondWaveData.ini) in the user config directory. This
information is then read and used to initialise the second wave options
in the difficulty screen.

Mods can also provide default on/off settings for second wave options in
the case where XComLW_SecondWaveData.ini doesn't exist. These are read
from XComLW_SecondWaveDataDefaults.ini.